make autoMaterializeAssetEvaluationId an ID field
BK

> Insert changelog entry or delete this section.

+++ python_modules/dagster-graphql/dagster_graphql/schema/asset_condition_evaluations.py
@@ -248,7 +248,7 @@ class GrapheneAutomationConditionEvaluationNode(graphene.ObjectType):
 
 class GrapheneAssetConditionEvaluationRecord(graphene.ObjectType):
     id = graphene.NonNull(graphene.ID)
-    evaluationId = graphene.NonNull(graphene.Int)
+    evaluationId = graphene.NonNull(graphene.BigInt)
     runIds = non_null_list(graphene.String)
     timestamp = graphene.NonNull(graphene.Float)
